pyproject.toml
[tool.poetry]
name = "webtronics_test"
version = "0.1.0"
description = ""
authors = ["Alexander Vidyakov <alex.vidyakov@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.1"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.0.0"
django-extensions = "^3.1.5"
psycopg2-binary = "^2.9.3"
drf-yasg = "^1.20.0"
furl = "^2.1.3"
pytest-django = "^4.5.2"
mimesis = "^5.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
